pyproject.toml
[tool.poetry]
name = "holiday-price-tracker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
selenium = "^4.26.1"
webdriver-manager = "^4.0.2"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
pre-commit = "^4.0.1"
commitizen = "^3.30.0"
requests = "^2.32.3"
aiogram = "^3.13.1"
bs4 = "^0.0.2"
python-dotenv = "^1.0.1"
playwright = "^1.48.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
